Refactor: project structure
