+++ src/app/(pages)/(myclublist)/myclublist/_apis/ClubListContentApi.ts
