+++ src/app/(pages)/(myclublist)/myclublist/_components/OneTimeClubCard.tsx
