+++ src/app/(pages)/(myclublist)/myclublist/_components/RegularClubCard.tsx
