+++ src/app/(pages)/(myclublist)/myclublist/page.tsx
@@ -1,4 +1,4 @@
-import ClubListContent from "@/components/myclublistcontent/ClubListContent";
+import ClubListContent from "@/app/(pages)/(myclublist)/myclublist/_components/ClubListContent";
 import { Suspense } from "react";
 
 export default function MyClubListPage() {
